[packaging] embed assets in binary, use as fallback when not on filesys
this applies to both assets and templates.
