+++ testrig/gin.go
@@ -28,8 +28,8 @@ import (
 // CreateGinTextContext creates a new gin.Context suitable for a test, with an instantiated gin.Engine.
 func CreateGinTestContext(rw http.ResponseWriter, r *http.Request) (*gin.Context, *gin.Engine) {
 	ctx, eng := gin.CreateTestContext(rw)
-	router.LoadTemplateFunctions(eng)
-	if err := router.LoadTemplates(eng); err != nil {
+	// FIXME @cleanup: why is this in addition to testrig/router.go#ConfigureTemplatesWithGin()?
+	if _, err := router.LoadTemplates(eng, ""); err != nil {
 		panic(err)
 	}
 	ctx.Request = r
